% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog.R
\name{blog_set_chunk}
\alias{blog_set_chunk}
\title{Blog setup}
\usage{
blog_set_chunk(message = FALSE, warning = FALSE, fig.showtext = TRUE,
  dev = "CairoPNG", fig.width = 11, fig.height = 7, dpi = 72)
}
\arguments{
\item{message}{= FALSE,}

\item{warning}{= FALSE,}

\item{fig.showtext}{= TRUE,}

\item{dev}{= "CairoPNG",}

\item{fig.width}{= 16,}

\item{fig.height}{= 9,}

\item{dpi}{= 72}
}
\description{
Blog setup
}

