% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{ccount}
\alias{ccount}
\alias{ccount_}
\title{Count the counts}
\usage{
ccount(data, ...)

ccount_(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables to group by, \code{count} arguments.}
}
\description{
Count the counts
}
\examples{
ccount(iris, Species)
ccount_(iris, "Species")
}
