% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{clean}
\alias{clean}
\title{clean data
This is a wrapper of some functions from janitor package}
\usage{
clean(data)
}
\arguments{
\item{data}{A data frame.}
}
\description{
clean data
This is a wrapper of some functions from janitor package
}
