% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_dbplyr.R
\name{compute2}
\alias{compute2}
\title{Force computation of a tbl_lazy in the database
Compute doesnt work well with T-SQL so this is a custom function to do the job}
\usage{
compute2(x, name, overwrite = FALSE)
}
\arguments{
\item{x}{A tbl_lazy.}

\item{name}{Name of temporary table on database.}

\item{overwrite}{A logical value indicating if remove the previous table with same name.}
}
\description{
Force computation of a tbl_lazy in the database
Compute doesnt work well with T-SQL so this is a custom function to do the job
}
