% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sekeleton_folders.R
\name{create_skeleton_folders}
\alias{create_skeleton_folders}
\title{Generate structure}
\usage{
create_skeleton_folders(dir = ".")
}
\arguments{
\item{dir}{Directory}
}
\description{
Generate structure
}

