% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_dbplyr.R
\name{dbRemoveTableIfExists}
\alias{dbRemoveTableIfExists}
\title{Remove table if exists}
\usage{
dbRemoveTableIfExists(con, name)
}
\arguments{
\item{con}{A connection object.}

\item{name}{A character string specifying a table name.}
}
\description{
Remove table if exists
}
