% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{filename_gen}
\alias{filename_gen}
\title{Auxiliar function to generate a filename following a pattern}
\usage{
filename_gen(folder = "plots", pattern = "plot", ext = "jpg",
  npad = 2)
}
\arguments{
\item{folder}{.}

\item{pattern}{.}

\item{ext}{.}

\item{npad}{.}
}
\description{
Auxiliar function to generate a filename following a pattern
}
