% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{ggsav}
\alias{ggsav}
\title{Function to save the last_plot using filename_gen}
\usage{
ggsav(folder = "plots", pattern = "plot", ext = "jpg", npad = 2,
  width = 16, height = 9, scale = 0.75)
}
\arguments{
\item{folder}{.}

\item{pattern}{.}

\item{ext}{.}

\item{npad}{.}

\item{width}{.}

\item{height}{.}

\item{scale}{.}
}
\description{
Function to save the last_plot using filename_gen
}

