% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giphy.R
\name{giphy}
\alias{giphy}
\title{Creating a md image and link source from a giphy url}
\usage{
giphy(url = "http://giphy.com/gifs/art-tech-algorithm-J7NAxMWW5Q6cg",
  txt_ttl = "giphy gif", txt_src = "source")
}
\arguments{
\item{url}{url}

\item{txt_ttl}{title}

\item{txt_src}{txt_src}
}
\description{
Creating a md image and link source from a giphy url
}

