% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog.R
\name{ico}
\alias{ico}
\title{Creating FA ico given name and color}
\usage{
ico(x = "tv", color = NULL)
}
\arguments{
\item{x}{Name of icon}

\item{color}{A color}
}
\description{
Creating FA ico given name and color
}
