% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictoc.R
\name{jbk_tic}
\alias{jbk_tic}
\alias{jbk_toc}
\alias{jbk_tictoc_clear}
\alias{jbk_tic_msg}
\alias{jbk_toc_msg}
\alias{jbk_tictoc_log}
\title{Tic Toe Wrappers}
\usage{
jbk_tic(msg = NULL, quiet = FALSE, func.tic = jbk_tic_msg)

jbk_toc(log = TRUE, func.toc = jbk_toc_msg)

jbk_tictoc_clear()

jbk_tic_msg(tic, msg = NULL)

jbk_toc_msg(tic, toc, msg, info)

jbk_tictoc_log()
}
\arguments{
\item{msg}{msg}

\item{quiet}{quiet}

\item{func.tic}{func.tic}

\item{log}{log}

\item{func.toc}{func.toc}

\item{tic}{tic}

\item{toc}{toc}

\item{info}{info}
}
\description{
Tic Toe Wrappers
}
\examples{

jbk_tictoc_clear()

jbk_tic(msg = "Go to sleep")

Sys.sleep(2)

jbk_toc()

jbk_tic(msg = "Go to sleep again!")

Sys.sleep(1)

jbk_toc()

jbk_tictoc_log()

}
