% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{list_rules_party}
\alias{list_rules_party}
\title{Get rules from partykit object}
\usage{
list_rules_party(x)
}
\arguments{
\item{x}{A party object.
library(partykit)
x <- ctree(Species ~ .,data = iris)
list_rules_party(x)}
}
\description{
Get rules from partykit object
}
