% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{sanitize_data}
\alias{sanitize_data}
\title{sanitize data
This is a wrapper of janitor package}
\usage{
sanitize_data(data)
}
\arguments{
\item{data}{A data frame.}
}
\description{
sanitize data
This is a wrapper of janitor package
}
