% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin_jekll_post.R
\name{spin_jekyll_post}
\alias{spin_jekyll_post}
\title{spin to md jekyll post}
\usage{
spin_jekyll_post(r_script)
}
\arguments{
\item{r_script}{r_script}
}
\description{
spin to md jekyll post
}

