% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{sqlFetch2}
\alias{sqlFetch2}
\title{RODBC::sqlFetch wrapper
Just return a tibble instead of a data.frame object and fixes the
\code{ambigous format} casting to character the datetime values}
\usage{
sqlFetch2(channel, sqtable)
}
\arguments{
\item{channel}{Connection handle returned by odbcConnect.}

\item{sqtable}{A database table name accessible from the connected DSN}
}
\description{
RODBC::sqlFetch wrapper
Just return a tibble instead of a data.frame object and fixes the
\code{ambigous format} casting to character the datetime values
}

