% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{sqlquery2}
\alias{sqlquery2}
\title{Execute a query given fields}
\usage{
sqlquery2(chn, table = "atable", fields = c("var1", "sum(var2)"))
}
\arguments{
\item{chn}{chn}

\item{table}{A table name}

\item{fields}{fields}
}
\description{
Execute a query given fields
}
