% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr.R
\name{str_clean}
\alias{str_clean}
\title{Transform name colums to human readable strings}
\usage{
str_clean(string, sep = "_|\\\\.", fun = stringr::str_to_title)
}
\arguments{
\item{string}{String}

\item{sep}{Pattern to make spaces strings}

\item{fun}{Function to apply before the separation}
}
\description{
Transform name colums to human readable strings
}
\examples{

string <- c("name.with...points", "or_Under___scores")
string

str_clean(string)
str_clean(string, fun = toupper)
str_clean(string, fun = stringr::str_to_title)

}
