% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog.R
\name{sw}
\alias{sw}
\title{Save widget for presentations which share same folder}
\usage{
sw(w, f)
}
\arguments{
\item{w}{Widget.}

\item{f}{File name.}
}
\description{
Save widget for presentations which share same folder
}
