% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_font.R
\name{theme_jbk}
\alias{theme_jbk}
\title{My ggplot2 go-to theme inspired by \code{hrbrmisc}}
\usage{
theme_jbk(base_family = "Lato", plot_title_family = "Open Sans Condensed",
  plot_title_face = "bold", striptext_family = plot_title_family,
  striptext_face = plot_title_face, strip_inverse = FALSE, base_size = 30)
}
\arguments{
\item{base_family}{base_family}

\item{plot_title_family}{plot_title_family}

\item{plot_title_face}{plot_title_face}

\item{striptext_family}{striptext_family}

\item{striptext_face}{striptext_face}

\item{base_size}{base_size}
}
\description{
My ggplot2 go-to theme inspired by \code{hrbrmisc}
}

