% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{theme_null}
\alias{theme_null}
\alias{tomorrow_night}
\title{My ggplot2 null theme}
\usage{
theme_null(...)

tomorrow_night(...)
}
\arguments{
\item{...}{Parameters}
}
\description{
My ggplot2 null theme
}
