% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressbar.R
\docType{data}
\name{TXTProgressBar}
\alias{TXTProgressBar}
\title{TXTProgressBar Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TXTProgressBar
}
\description{
A new progresss bar with text messages
}
\details{
set.seed(123)
x <- sample(fruit, 6)
mpb <- TXTProgressBar$new(x)
for(y in x){

Sys.sleep(0.5)
mpb$update(y)
}
}
\keyword{datasets}
