% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_importance.R
\name{var_importance}
\alias{var_importance}
\title{Importance variable for predictive methods}
\usage{
var_importance(x)
}
\arguments{
\item{x}{A RandomForest or RRF object.}
}
\description{
Importance variable for predictive methods
}
