% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{wf_create_folders}
\alias{wf_create_folders}
\title{Generate structure}
\usage{
wf_create_folders(dir = ".", folders = c("R", "data", "outputs"))
}
\arguments{
\item{dir}{Directory}

\item{folders}{Folders to create in the dir directory}
}
\description{
Generate structure
}
