% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{clean}
\alias{clean}
\title{clean data
This is a wrapper of some functions from janitor package}
\usage{
clean(data, integer64_to_num = TRUE, factor_to_char = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{integer64_to_num}{Logical to transform integer64 to numeric}

\item{factor_to_char}{Logical to transform factor to characters}
}
\description{
clean data
This is a wrapper of some functions from janitor package
}
