% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_dbplyr.R
\name{countp}
\alias{countp}
\title{Add percent to \code{count}}
\usage{
countp(data, ..., sort = FALSE, add_pcum = TRUE, ungroup = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables to group by, \code{count} arguments.}

\item{sort}{same sort argument.}

\item{add_pcum}{Add or not cumulative percent column.}

\item{ungroup}{Remove the group before calculate the percents.}
}
\description{
Add percent to \code{count}
}
\examples{
countp(mtcars, cyl)
countp(mtcars, cyl, am)
countp(mtcars, cyl, am, ungroup = FALSE)
countp(iris, "Species")
}
