% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{format_tbl}
\alias{format_tbl}
\title{Format table to report}
\usage{
format_tbl(data, digits = 2, NA_symbol = "-", fun_title = str_clean)
}
\arguments{
\item{data}{data frame}

\item{digits}{Same argument as round}

\item{NA_symbol}{NA_symbol}

\item{fun_title}{fun_title}
}
\description{
Format table to report
}
\examples{

data <- data.frame(
  a_value.e = c(NA, runif(3)),
  `another.vAlue__here` = c(NA, runif(3)) * 10000
)

data
format_tbl(data)
format_tbl(data, digits = 0, NA_symbol = "***", fun_title = toupper)

}
