% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giphy.R
\name{giphy}
\alias{giphy}
\title{Creating a md image and link source from a giphy url}
\usage{
giphy(id = "RgfGmnVvt8Pfy", txt_ttl = "giphy gif", addsource = TRUE)
}
\arguments{
\item{id}{id}

\item{txt_ttl}{title}

\item{addsource}{addsource}
}
\description{
Creating a md image and link source from a giphy url
}

