% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_RODBC.R
\name{sqlquery}
\alias{sqlquery}
\title{My ggplot2 go-to theme inspired by \code{hrbrmisc}}
\usage{
sqlquery(...)
}
\arguments{
\item{...}{Same arguments as \code{RODBC::sqlQuery}}
}
\description{
Just return a tibble instead of a data.frame object.
}

