% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{sqlQuery}
\alias{sqlQuery}
\title{RODBC::sqlQuery wrapper
Just return a tibble instead of a data.frame object.}
\usage{
sqlQuery(...)
}
\arguments{
\item{...}{Same arguments as \code{RODBC::sqlQuery}}
}
\description{
RODBC::sqlQuery wrapper
Just return a tibble instead of a data.frame object.
}

