% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{sqlQuery2}
\alias{sqlQuery2}
\title{RODBC sqlQuery wrapper
Just return a tibble instead of a data.frame object.}
\usage{
sqlQuery2(channel, query)
}
\arguments{
\item{channel}{Connection handle as returned by odbcConnect.}

\item{query}{Any valid SQL statement.}
}
\description{
RODBC sqlQuery wrapper
Just return a tibble instead of a data.frame object.
}
