% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{theme_jbk}
\alias{theme_jbk}
\title{My ggplot2 go-to theme inspired by \code{hrbrthemes}}
\usage{
theme_jbk(base_family = "", ...)
}
\arguments{
\item{base_family}{base_family}

\item{...}{Parameters for \code{hrbrthemes::theme_ipsum}}
}
\description{
My ggplot2 go-to theme inspired by \code{hrbrthemes}
}
\examples{


\dontrun{
require(ggplot2)
require(hrbrthemes)
require(extrafont)
loadfonts()

ggplot(mtcars) +
  geom_point(aes(mpg, hp, color = factor(carb))) +
  scale_color_ipsum() +
  labs(x="Fuel effiiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
 theme_jbk()
}

}
